Remove pkgindb_simple_callback().
At best it was useless.  If used to check that a SQL statement has been
executed successfully then a NULL callback would achieve the same thing
faster with less code.

At worst it was confusing.  The argv == NULL check is as far as I can
tell never going to be true (if there are no results returned then the
callback would not be called), so the callback will always return true.

This matters for the current COMPAT_CHECK test, as the query is written
to test not that the query is valid, but that it returns valid results.
As pkgindb_simple_callback() did not check that argv[0] was not NULL,
then the check will succeed even if no results were returned.

The upgrade tests will be updated in due course to be more explicit and
differentiate between an incompatible database and one that has not yet
been initialised.